requirements.txt
streamlit
crewai
langchain_groq
langchain.agents
